@@ -8,10 +8,10 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.24.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.0.0-beta.6",
     "react-dev-utils": "^11.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-router": "^5.2.1",
     "react-router-bootstrap": "^0.26.0",
     "react-router-dom": "^6.2.1",
@@ -30,18 +30,9 @@
       "react-app/jest"
     ]
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
+  "browserslist": [
+    "defaults"
+  ],
   "devDependencies": {
     "base": "^3.0.0",
     "browserslist": "^4.17.3",
